Consolidated plot_heatmap: tested working with results from EPIC and ERC
